// ui/src/lib/booking-hotel-api/model/order-save-request.ts
/**
 * main-app API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderSaveRequest { 
    id?: number;
    fromDate: string;
    toData: string;
    hotelId: number;
    userId: number;
    roomTypeId: number;
}